src/azure-ai-102/information-extraction/05-rag-pipeline/pyproject.toml
[project]
name = "05-rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-ai-contentunderstanding>=1.0.1",
    "azure-search-documents==11.6.0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
]
